refactor(main.go): Splitting main.go to mutliple files.
