--- pkg/mshproto/meshprotocol.go
@@ -1,9 +1,0 @@
-// Constants for the Meshtastic protocol
-package meshtastic
-
-const (
-	START1          = 0x94
-	START2          = 0xC3
-	HEADER_LEN      = 4
-	MAX_PACKET_SIZE = 512
-)
